Initialize the app’s basic data models
Create the User, Task and TaskNote structs. The User struct represents users of the application. A user should register at the application to create tasks. An authenticated user can add tasks, which will be represented with the Task struct. A user can add notes against each task, which will be represented with the TaskNote struct. The TaskNote entity holds the child details of its parent entity: Task.